refactor: move messaging provider
